backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
openai==1.35.0
pandas==2.2.2
pymupdf==1.24.5
reportlab==4.2.2
python-dotenv==1.0.1
werkzeug==3.0.3